// frontend/src/Interfaces/LecturerProfileModel.ts
export interface LecturerProfileModel {
 id: string | null;               
  personalID: string | null;
    email: string;
    userName: string;
    lastName: string;
    gender: string | null;
    birthdate: string | null;
    departmentId: string | null;
    cityId: string | null;
    address: string | null;
    personalEmail: string | null;
    phoneNumber: string | null;
  academicGrade: string;
  
}